+++ pytest_services/service.py
@@ -15,15 +15,15 @@ WRONG_FILE_NAME_CHARS_RE = re.compile(r'[^\w_-]')
 
 
 @pytest.fixture(scope='session')
-def run_services(request, slave_id):
+def run_services(request, worker_id):
     """Indicate whether the services should run or not."""
-    return slave_id != 'local' or request.config.option.run_services
+    return worker_id != 'local' or request.config.option.run_services
 
 
 @pytest.fixture(scope='session')
-def slave_id(request):
+def worker_id(request):
     """
-    The id of the slave if tests are run using xdist.
+    The id of the worker if tests are run using xdist.
 
     It is set to `'local'` if tests are not run using xdist.
 
@@ -31,20 +31,27 @@ def slave_id(request):
     that belong to different test sessions.
 
     """
-    return WRONG_FILE_NAME_CHARS_RE.sub('_', getattr(request.config, 'slaveinput', {}).get('slaveid', 'local'))
+    return WRONG_FILE_NAME_CHARS_RE.sub('_', getattr(request.config, 'workerinput', {}).get('workerid', 'local'))
 
 
 @pytest.fixture(scope='session')
-def session_id(request, slave_id, run_services):
+def slave_id(request, worker_id):
+    msg = "The `slave_id` fixture is deprecated; use `worker_id` instead."
+    warnings.warn(msg, DeprecationWarning)
+    return worker_id
+
+
+@pytest.fixture(scope='session')
+def session_id(request, worker_id, run_services):
     """The test session id.
 
     It is supposed to be globally unique.
 
     """
     # UUID should be enough, other fields are added for the debugging purposes.
-    session_id = '{random}-{slave_id}'.format(
+    session_id = '{random}-{worker_id}'.format(
         random=uuid.uuid4().hex,
-        slave_id=slave_id,
+        worker_id=worker_id,
     )
 
     return session_id